remove confusing deserialize(inputStream)
setInputStream(), which is called by setInput(), rewinds the stream, so
the deserialize() function can only deserialize the first object on the
stream. That's why the open-stream and ObjectInput stuff was needed.

removing deserialize(inputStream) because a user may think that it can
deserialize more than one object from the stream.